requirements.txt
requests==2.32.4
fastapi==0.116.1
uvicorn==0.35.0
python-telegram-bot==22.3
python-decouple==3.8
async-lru==2.0.5
google-genai==1.28.0